Create common.rs for shared api core functions
